Split integration test code
