+++ tests/test_integration.py
@@ -112,7 +112,6 @@ def kpet_run_generate(db_name, *args):
                         "-k", "kernel.tar.gz", "-a", "arch", *args)
 
 
-# pylint: disable=too-many-public-methods
 class IntegrationTests(unittest.TestCase):
     """Integration tests"""
 
@@ -155,6 +154,7 @@ class IntegrationTests(unittest.TestCase):
         if errors:
             raise AssertionError("\n".join(errors))
 
+
     def assertKpetSrcMatchesTwoSuites(self, db_name):
         """
         Assert kpet source-matches two suites properly.
@@ -190,6 +190,7 @@ class IntegrationTests(unittest.TestCase):
             get_patch_path("misc/files_ghi.diff"),
             stdout_matching=r'.*<job>\s*</job>.*')
 
+
     def assertKpetSrcMatchesOneOfTwoSuites(self, db_name):
         """
         Assert kpet source-matches one of two suites properly.
@@ -223,6 +224,7 @@ class IntegrationTests(unittest.TestCase):
             get_patch_path("misc/files_ghi.diff"),
             stdout_matching=r'.*<job>\s*</job>.*')
 
+
     def assertKpetSrcMatchesNoneOfTwoSuites(self, db_name):
         """
         Assert kpet source-matches none of two suites properly.
@@ -241,513 +243,3 @@ class IntegrationTests(unittest.TestCase):
             get_patch_path("misc/files_def.diff"),
             get_patch_path("misc/files_ghi.diff"),
             stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_empty_tree_list(self):
-        """Test tree listing with empty database"""
-        self.assertKpetProduces(kpet_with_db, "empty/db", "tree", "list")
-
-    def test_empty_run_generate(self):
-        """Test run generation with empty database"""
-        self.assertKpetProduces(
-            kpet_run_generate, "empty/db",
-            status=1,
-            stderr_matching=r'.*Architecture "arch" not found.*')
-
-    def test_minimal_run_generate(self):
-        """Test run generation with empty database"""
-        self.assertKpetProduces(kpet_run_generate, "minimal",
-                                stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_missing_tree_template_run_generate(self):
-        """Test run generation with a missing tree template"""
-        self.assertKpetProduces(kpet_run_generate,
-                                "invalid/semantics/missing_tree_template",
-                                "-t", "missing_template",
-                                status=1,
-                                stderr_matching=r'.*TemplateNotFound.*')
-
-    def test_missing_suite_file_run_generate(self):
-        """Test run generation with a missing suite file"""
-        self.assertKpetProduces(kpet_run_generate,
-                                "invalid/semantics/missing_suite_file",
-                                status=1,
-                                stderr_matching=r'.*missing.yaml.*')
-
-    def test_invalid_top_yaml_tree_list(self):
-        """Test tree listing with invalid YAML in the top database file"""
-        self.assertKpetProduces(kpet_with_db, "invalid/yaml/top",
-                                "tree", "list",
-                                status=1,
-                                stderr_matching=r'.*yaml.parser.ParserError.*')
-
-    def test_invalid_suite_yaml_tree_list(self):
-        """Test tree listing with invalid YAML in a suite file"""
-        self.assertKpetProduces(kpet_with_db, "invalid/yaml/suite",
-                                "tree", "list",
-                                status=1,
-                                stderr_matching=r'.*yaml.parser.ParserError.*')
-
-    def test_invalid_top_data_tree_list(self):
-        """Test tree listing with invalid data in the top database file"""
-        self.assertKpetProduces(kpet_with_db, "invalid/data/top",
-                                "tree", "list",
-                                status=1,
-                                stderr_matching=r'.*Invalid Base data.*')
-
-    def test_invalid_suite_data_tree_list(self):
-        """Test tree listing with invalid data in a suite file"""
-        self.assertKpetProduces(kpet_with_db, "invalid/data/suite",
-                                "tree", "list",
-                                status=1,
-                                stderr_matching=r'.*Invalid Suite data.*')
-
-    def test_empty_suite_run_generate(self):
-        """Test run generation with an empty suite"""
-        self.assertKpetProduces(kpet_run_generate, "empty/suite",
-                                stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_empty_case_no_patterns_run_generate(self):
-        """Test run generation with an empty test case without patterns"""
-        self.assertKpetProduces(
-            kpet_run_generate, "empty/case_no_patterns",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-
-    def test_empty_case_with_a_pattern_run_generate(self):
-        """Test run generation with an empty test case with a pattern"""
-        self.assertKpetProduces(
-            kpet_run_generate, "empty/case_with_a_pattern",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-
-    def test_match_sources_one_case_no_patterns(self):
-        """Test source-matching a case with no patterns"""
-        # Matches baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/no_patterns",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/no_patterns",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-
-    def test_match_sources_one_case_one_pattern(self):
-        """Test source-matching a case with one pattern"""
-        # Matches baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/one_pattern",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches patches it should
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/one_pattern",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Doesn't match patches it shouldn't
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/one_pattern",
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_sources_one_case_two_patterns(self):
-        """Test source-matching a case with two patterns"""
-        # Matches baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/two_patterns",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches first patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/two_patterns",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches second patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/two_patterns",
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches both patches only once
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/two_patterns",
-            get_patch_path("misc/files_abc.diff"),
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Doesn't match patches it shouldn't
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/one_case/two_patterns",
-            get_patch_path("misc/files_ghi.diff"),
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_sources_two_cases(self):
-        """Test source-matching two cases"""
-        # Both match baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_cases",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*'
-                            r'case1\s*case2\s*</job>.*')
-        # First matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_cases",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Second matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_cases",
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case2\s*</job>.*')
-        # Both match their patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_cases",
-            get_patch_path("misc/files_abc.diff"),
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*'
-                            r'case1\s*case2\s*</job>.*')
-        # None match other patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_cases",
-            get_patch_path("misc/files_ghi.diff"),
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_sources_two_suites(self):
-        """Test source-matching two suites"""
-        # Both match baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_suites",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
-        # First matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_suites",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Second matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_suites",
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite2\s*case2\s*</job>.*')
-        # Both match their patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_suites",
-            get_patch_path("misc/files_abc.diff"),
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
-        # None match other patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/two_suites",
-            get_patch_path("misc/files_ghi.diff"),
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_sources_specific_suite(self):
-        """Test source-matching with a specific suite"""
-        # Only non-specific suite matches baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/suite",
-            stdout_matching=r'.*<job>\s*HOST\s*suite2\s*case2\s*</job>.*')
-        # First matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/suite",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Second matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/suite",
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite2\s*case2\s*</job>.*')
-        # All suites can match if provided appropriate patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/suite",
-            get_patch_path("misc/files_abc.diff"),
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
-        # None match other patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/suite",
-            get_patch_path("misc/files_ghi.diff"),
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_sources_specific_case(self):
-        """Test source-matching with a specific case"""
-        # Only non-specific case matches baseline
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/case",
-            stdout_matching=r'.*<job>\s*HOST\s*suite2\s*case2\s*</job>.*')
-        # First matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/case",
-            get_patch_path("misc/files_abc.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Second matches its patch
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/case",
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite2\s*case2\s*</job>.*')
-        # All cases can match if provided appropriate patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/case",
-            get_patch_path("misc/files_abc.diff"),
-            get_patch_path("misc/files_def.diff"),
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
-        # None match other patches
-        self.assertKpetProduces(
-            kpet_run_generate, "match/sources/specific/case",
-            get_patch_path("misc/files_ghi.diff"),
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_arches_no_patterns(self):
-        """Test architecture-matching a case with no patterns"""
-        # Doesn't match a non-empty (default "arch") architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/no_patterns",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match empty architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/no_patterns", "-a", "",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_arches_one_pattern(self):
-        """Test architecture-matching a case with one pattern"""
-        # Matches default ("arch") architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/one_pattern",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Doesn't match empty architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/one_pattern", "-a", "",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match another architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/one_pattern", "-a", "not_arch",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_arches_two_patterns(self):
-        """Test architecture-matching a case with two patterns"""
-        # Matches default ("arch") architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/two_patterns",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches non-default (but listed) architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/two_patterns",
-            "-a", "other_arch",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Doesn't match empty architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/two_patterns", "-a", "",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match another architecture
-        self.assertKpetProduces(
-            kpet_run_generate, "match/arches/two_patterns", "-a", "not_arch",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_trees_no_patterns(self):
-        """Test tree-matching a case with no patterns"""
-        # Doesn't match a non-empty (default "tree") tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/no_patterns",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match empty tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/no_patterns", "-t", "",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_trees_one_pattern(self):
-        """Test tree-matching a case with one pattern"""
-        # Matches default ("tree") tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/one_pattern",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Doesn't match empty tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/one_pattern", "-t", "",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match another tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/one_pattern", "-t", "not_tree",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_trees_two_patterns(self):
-        """Test tree-matching a case with two patterns"""
-        # Matches default ("tree") tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/two_patterns",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Matches non-default (but listed) tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/two_patterns",
-            "-t", "other_tree",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*</job>.*')
-        # Doesn't match empty tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/two_patterns", "-t", "",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match another tree
-        self.assertKpetProduces(
-            kpet_run_generate, "match/trees/two_patterns", "-t", "not_tree",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_location_types_no_patterns(self):
-        """Test location-type matching a case with no patterns"""
-        # Doesn't match a tarball path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/no_patterns",
-            "-k", "kernel.tar.gz",
-            stdout_matching=r'.*<job>\s*</job>.*')
-        # Doesn't match an RPM URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/no_patterns",
-            "-k", "http://example.com/kernel.rpm",
-            stdout_matching=r'.*<job>\s*</job>.*')
-
-    def test_match_location_types_one_pattern(self):
-        """Test location-type matching cases with one pattern"""
-        # Only one case matches a tarball path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/one_pattern",
-            "-k", "kernel.tar.gz",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*'
-                            r'tarball-path\s*</job>.*')
-        # Only one case matches a tarball URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/one_pattern",
-            "-k", "http://example.com/kernel.tar.gz",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*'
-                            r'tarball-url\s*</job>.*')
-        # Only one case matches an RPM path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/one_pattern",
-            "-k", "kernel.rpm",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*rpm-path\s*</job>.*')
-        # Only one case matches an RPM URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/one_pattern",
-            "-k", "http://example.com/kernel.rpm",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*rpm-url\s*</job>.*')
-        # Only one case matches a repo path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/one_pattern",
-            "-k", "repo",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*repo-path\s*</job>.*')
-        # Only one case matches a repo URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/one_pattern",
-            "-k", "http://example.com/repo",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*repo-url\s*</job>.*')
-
-    def test_match_location_types_two_patterns(self):
-        """Test location-type matching cases with two patterns"""
-        # Tarball case matches a tarball path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/two_patterns",
-            "-k", "kernel.tar.gz",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*tarball\s*</job>.*')
-        # Tarball case matches a tarball URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/two_patterns",
-            "-k", "http://example.com/kernel.tar.gz",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*tarball\s*</job>.*')
-        # RPM case matches an RPM path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/two_patterns",
-            "-k", "kernel.rpm",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*rpm\s*</job>.*')
-        # RPM case matches an RPM URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/two_patterns",
-            "-k", "http://example.com/kernel.rpm",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*rpm\s*</job>.*')
-        # Repo case matches a repo path
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/two_patterns",
-            "-k", "repo",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*repo\s*</job>.*')
-        # Repo case matches a repo URL
-        self.assertKpetProduces(
-            kpet_run_generate, "match/location_types/two_patterns",
-            "-k", "http://example.com/repo",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*repo\s*</job>.*')
-
-    def test_multihost_no_types_no_regex_no_suites(self):
-        """Test multihost support without types/regex/suites"""
-        self.assertKpetSrcMatchesNoneOfTwoSuites(
-            "multihost/no_types/no_regex/no_suites")
-
-    def test_multihost_no_types_no_regex_two_suites(self):
-        """Test multihost support without types/regex and two suites"""
-        self.assertKpetSrcMatchesTwoSuites(
-            "multihost/no_types/no_regex/two_suites")
-
-    def test_multihost_no_types_wildcard_regex_two_suites(self):
-        """
-        Test multihost support without types, with a DB-level wildcard regex,
-        and two suites.
-        """
-        self.assertKpetSrcMatchesTwoSuites("multihost/no_types/db_regex")
-
-    def test_multihost_one_type_no_regex_no_suites(self):
-        """Test multihost support with one type, but no regex/suites"""
-        self.assertKpetSrcMatchesNoneOfTwoSuites(
-            "multihost/one_type/no_regex/no_suites")
-
-    def test_multihost_one_type_no_regex_two_suites(self):
-        """Test multihost support with one type, no regex, and two suites"""
-        self.assertKpetSrcMatchesNoneOfTwoSuites(
-            "multihost/one_type/no_regex/two_suites")
-
-    def test_multihost_one_type_db_regex(self):
-        """Test multihost support with one type and DB-level regex"""
-        self.assertKpetSrcMatchesTwoSuites("multihost/one_type/db_regex")
-
-    def test_multihost_one_type_suite_regex(self):
-        """Test multihost support with one type and suite-level regexes"""
-        self.assertKpetSrcMatchesOneOfTwoSuites(
-            "multihost/one_type/suite_regex")
-
-    def test_multihost_one_type_case_regex(self):
-        """Test multihost support with one type and case-level regexes"""
-        self.assertKpetSrcMatchesOneOfTwoSuites(
-            "multihost/one_type/case_regex")
-
-    def test_multihost_one_type_wrong_regex(self):
-        """Test multihost support with one type a wrong DB-level regex"""
-        self.assertKpetSrcMatchesNoneOfTwoSuites(
-            "multihost/one_type/wrong_regex")
-
-    def test_multihost_two_types_one_case_each(self):
-        """Test multihost support with two types matching one case each"""
-        self.assertKpetProduces(
-            kpet_run_generate, "multihost/two_types/one_case_each",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'HOST\s*suite2\s*case2\s*</job>.*')
-
-    def test_multihost_two_types_both_cases_first(self):
-        """
-        Test multihost support with two types and both cases matching the
-        first one.
-        """
-        # TODO Distinguish host types somehow
-        self.assertKpetProduces(
-            kpet_run_generate, "multihost/two_types/both_cases_first",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
-
-    def test_multihost_two_types_both_cases_second(self):
-        """
-        Test multihost support with two types and both cases matching the
-        second one.
-        """
-        # TODO Distinguish host types somehow
-        self.assertKpetProduces(
-            kpet_run_generate, "multihost/two_types/both_cases_second",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
-
-    def test_multihost_two_types_both_cases_both(self):
-        """
-        Test multihost support with two types and both cases matching both
-        types.
-        """
-        # TODO Distinguish host types somehow
-        self.assertKpetProduces(
-            kpet_run_generate, "multihost/two_types/both_cases_both",
-            stdout_matching=r'.*<job>\s*HOST\s*suite1\s*case1\s*'
-                            r'suite2\s*case2\s*</job>.*')
